Changed S3EndpointConfiguration to static inner S3Endpoint.Configuration class
